% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIPackage.R
\name{APIpackage}
\alias{APIpackage}
\alias{create_database}
\title{API to get data from Swedish Parliment open Data site}
\usage{
create_database(appointment, person_data_frame, csv_file)
}
\arguments{
\item{appointment}{member appointments}

\item{person_data_frame}{member information}

\item{csv_file}{voting information of members}
}
\value{
it will fetch the data from API URL, parse it according the requirement and send to create_database function to insert into database

it is create the database and insert the data passed as arguments into tables
}
\description{
API to get data from Swedish Parliment open Data site
}
\examples{
\dontrun{
data_fetch()
}

\dontrun{
create_database(appointment,person_data_frame,csv_file)
}
}
